% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_moments.R
\name{binomial_moments}
\alias{binomial_moments}
\title{Moments of the Binomial Distribution B(n,p)}
\usage{
binomial_moments(n, p)
}
\arguments{
\item{n}{Integer number of trials (must be >= 0).}

\item{p}{Probability of success (must be between 0 and 1).}
}
\value{
A list with elements 'mean' and 'variance'.
}
\description{
Compute the mean and variance for the binomial
distribution with parameters n and p.
}
\examples{
binomial_moments(10, 0.5)

}
