% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{binomial_pmf}
\alias{binomial_pmf}
\alias{binomial_cdf}
\alias{binomial_rnd}
\title{Binomial Distribution B(n,p)}
\usage{
binomial_pmf(k, n, p)

binomial_cdf(q, n, p)

binomial_rnd(num_samples, n, p)
}
\arguments{
\item{k}{Integer value(s) for PMF (x values). looking for k}

\item{n}{Integer number of trials (must be >= 0). how many times i do the experiment}

\item{p}{Probability of success (must be between 0 and 1). probability}

\item{q}{Numeric value(s) for CDF. x <= k}

\item{num_samples}{Number of random observations to generate.}
}
\value{
\code{binomial_pmf} gives the probability mass.
\code{binomial_cdf} gives the cumulative distribution function.
\code{binomial_rnd} generates random deviates.
}
\description{
PMF, CDF, and RNG for a binomial distribution with parameters n and p.
}
\details{
The Binomial distribution models the number of successes in n independent Bernoulli trials
with probability of success p. Support is \{0, 1, ..., n\}.
}
\seealso{
Other discrete-distributions: 
\code{\link{discrete_uniform_pmf}()},
\code{\link{poisson_pmf}()}
}
\concept{discrete-distributions}
