% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi-square_moments.R
\name{chi_square_moments}
\alias{chi_square_moments}
\title{Moments of the Chi-Square Distribution}
\usage{
chi_square_moments(df)
}
\arguments{
\item{df}{Degrees of freedom (n).}
}
\value{
A list with elements 'mean' and 'variance'.
}
\description{
Compute the mean and variance for the Chi-Square distribution.
}
