% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_uniform_moments.R
\name{continuous_uniform_moments}
\alias{continuous_uniform_moments}
\title{Moments of the Continuous Uniform Distribution U(a,b)}
\usage{
continuous_uniform_moments(min = 0, max = 1)
}
\arguments{
\item{min}{Lower limit (a).}

\item{max}{Upper limit (b).}
}
\value{
A list with elements 'mean' and 'variance'.
}
\description{
Compute the mean and variance for the continuous uniform distribution.
}
