% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_uniform.R
\name{continuous_uniform_pdf}
\alias{continuous_uniform_pdf}
\alias{continuous_uniform_cdf}
\alias{continuous_uniform_rnd}
\title{Continuous Uniform Distribution U(a,b)}
\usage{
continuous_uniform_pdf(x, min = 0, max = 1)

continuous_uniform_cdf(q, min = 0, max = 1)

continuous_uniform_rnd(n, min = 0, max = 1)
}
\arguments{
\item{x}{Numeric value(s) for PDF (density).}

\item{min}{Lower limit of the distribution (a).}

\item{max}{Upper limit of the distribution (b).}

\item{q}{Numeric value(s) for CDF.}

\item{n}{Number of random observations to generate.}
}
\value{
\code{continuous_uniform_pdf} gives the density.
\code{continuous_uniform_cdf} gives the distribution function.
\code{continuous_uniform_rnd} generates random deviates.
}
\description{
PDF, CDF, and RNG for a continuous uniform distribution on the interval \code{[a, b]}.
}
\details{
The continuous uniform distribution describes a variable where all values
within the interval \code{[a, b]} are equally probable.
Support is the interval \code{[a, b]}.
}
\seealso{
Other continuous-distributions: 
\code{\link{chi_square_pdf}()},
\code{\link{fisher_f_pdf}()},
\code{\link{normal_pdf}()},
\code{\link{student_t_pdf}()}
}
\concept{continuous-distributions}
