% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_f_moments.R
\name{fisher_f_moments}
\alias{fisher_f_moments}
\title{Moments of the Fisher-Snedecor F-Distribution}
\usage{
fisher_f_moments(df1, df2)
}
\arguments{
\item{df1}{Degrees of freedom of the numerator (m).}

\item{df2}{Degrees of freedom of the denominator (n).}
}
\value{
A list with elements 'mean' and 'variance'.
}
\description{
Compute the mean and variance for the F-distribution.
}
