% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_moments.R
\name{geometric_moments}
\alias{geometric_moments}
\title{Moments of the Geometric Distribution Geom(p)}
\usage{
geometric_moments(p)
}
\arguments{
\item{p}{Probability of success (0 < p <= 1).}
}
\value{
A list with elements 'mean' and 'variance'.
}
\description{
Compute the mean and variance for the Geometric distribution.
defined as the number of trials to achieve the first success.
}
