% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_moments.R
\name{normal_moments}
\alias{normal_moments}
\title{Moments of the Normal Distribution N(mu, sigma)}
\usage{
normal_moments(mean = 0, sd = 1)
}
\arguments{
\item{mean}{The mean (mu).}

\item{sd}{The standard deviation (sigma > 0).}
}
\value{
A list with elements 'mean' and 'variance'.
}
\description{
Compute the mean and variance for the normal distribution.
}
