% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{normal_pdf}
\alias{normal_pdf}
\alias{normal_cdf}
\alias{normal_rnd}
\title{Normal Distribution N(mu, sigma)}
\usage{
normal_pdf(x, mean = 0, sd = 1)

normal_cdf(q, mean = 0, sd = 1)

normal_rnd(n, mean = 0, sd = 1)
}
\arguments{
\item{x}{Numeric value(s) for PDF (density).}

\item{mean}{The mean (mu) of the distribution (default = 0).}

\item{sd}{The standard deviation (sigma) of the distribution (must be > 0, default = 1).}

\item{q}{Numeric value(s) for CDF.}

\item{n}{Number of random observations to generate.}
}
\value{
\code{normal_pdf} gives the density.
\code{normal_cdf} gives the distribution function.
\code{normal_rnd} generates random deviates.
}
\description{
PDF, CDF, and RNG for a normal distribution with mean mu and standard deviation sigma.
}
\details{
The normal distribution (Gaussian distribution) is defined by its mean and standard deviation.
The standard normal distribution occurs when mu = 0 and sigma = 1.
Support is the set of all real numbers \code{(-Inf, Inf)}.
}
\examples{
normal_pdf(0, mean = 0, sd = 1)
normal_cdf(1.96, mean = 0, sd = 1)

}
\seealso{
Other continuous-distributions: 
\code{\link{chi_square_pdf}()},
\code{\link{continuous_uniform_pdf}()}
}
\concept{continuous-distributions}
