% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_moments.R
\name{poisson_moments}
\alias{poisson_moments}
\title{Moments of the Poisson Distribution P(λ)}
\usage{
poisson_moments(lambda)
}
\arguments{
\item{lambda}{Positive rate parameter.}
}
\value{
A list with elements \code{mean} and \code{variance}.
}
\description{
Compute the mean and variance of a Poisson distribution with parameter λ.
}
\examples{
poisson_moments(3)
}
