% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spsqr_and_f.R
\name{pooled_variance}
\alias{pooled_variance}
\title{Two Sample Statistics Helper Functions}
\usage{
pooled_variance(A, B)
}
\arguments{
\item{A}{Numeric vector for the first group.}

\item{B}{Numeric vector for the second group.}
}
\value{
Numeric value of the calculated statistic.
}
\description{
Functions to calculate Pooled Variance and Welch's Degrees of Freedom
based on the formulas provided in Chapter 6 (Page 21/30).
}
