% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-t_moments.R
\name{student_t_moments}
\alias{student_t_moments}
\title{Moments of the Student's t-Distribution}
\usage{
student_t_moments(df)
}
\arguments{
\item{df}{Degrees of freedom (n).}
}
\value{
A list with elements 'mean' and 'variance'.
}
\description{
Compute the mean and variance for the Student's t-distribution.
}
