% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-t.R
\name{student_t_pdf}
\alias{student_t_pdf}
\alias{student_t_cdf}
\alias{student_t_rnd}
\title{Student's t-Distribution T(df)}
\usage{
student_t_pdf(x, df)

student_t_cdf(q, df)

student_t_rnd(n, df)
}
\arguments{
\item{x}{Numeric value(s) for PDF.}

\item{df}{Degrees of freedom (must be > 0).}

\item{q}{Numeric value(s) for CDF.}

\item{n}{Number of random observations to generate.}
}
\value{
\code{student_t_pdf} gives the density.
\code{student_t_cdf} gives the distribution function.
\code{student_t_rnd} generates random deviates.
}
\description{
PDF, CDF, and RNG for the Student's t-distribution with df degrees of freedom.
}
\details{
The Student's t-distribution is symmetric and bell-shaped, like the normal distribution,
but has heavier tails. It is used extensively in hypothesis testing and confidence intervals
when the population variance is unknown.
Support is the set of all real numbers \code{(-Inf, Inf)}.
}
\seealso{
Other continuous-distributions: 
\code{\link{chi_square_pdf}()},
\code{\link{continuous_uniform_pdf}()},
\code{\link{normal_pdf}()}
}
\concept{continuous-distributions}
